requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dominate==2.6.0
flake8==3.8.4
Flask==1.1.2
Flask-Bootstrap4==4.0.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
requests==2.25.0
soupsieve==2.0.1
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
